int to float

+++ main.py
@@ -1,4 +1,4 @@
 intTemp = input("Please input 'C temp")
-intTemp = int(intTemp)
+intTemp = float(intTemp)
 intOutput = ((intTemp * 9 )/5) + 32
 print(intOutput)
